// types.ts
export interface FormData {
  fullName: string;
  companyName: string;
  phone: string;
  email: string;
  website: string;
  niche: string;
  objectives: string[];
  otherObjective: string;
  challenge: string;
  targetAudience: string;
  competitors: string;
  services: string[];
  budget: string;
  deadline: string;
  foundUsBy: string;
  additionalInfo: string;
}